# packages/api-spec/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/auth/facebook": {
      "post": {
        "operationId": "CitizenAuthController_facebook",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FacebookLoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Citizen Authentication"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "CitizenAuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Citizen Authentication"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "CitizenAuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Citizen Authentication"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "CitizenAuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Citizen Authentication"
        ]
      }
    },
    "/api/v1/admin/auth/login": {
      "post": {
        "operationId": "AdminAuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminLoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Admin Authentication"
        ]
      }
    },
    "/api/v1/admin/auth/refresh": {
      "post": {
        "operationId": "AdminAuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Admin Authentication"
        ]
      }
    },
    "/api/v1/admin/auth/logout": {
      "post": {
        "operationId": "AdminAuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Admin Authentication"
        ]
      }
    },
    "/api/v1/admin/auth/me": {
      "get": {
        "operationId": "AdminAuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin Authentication"
        ]
      }
    },
    "/api/v1/admin/auth/change-password": {
      "patch": {
        "operationId": "AdminAuthController_change",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin Authentication"
        ]
      }
    },
    "/api/v1/incidents": {
      "post": {
        "operationId": "CitizenIncidentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIncidentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Citizen Incidents"
        ]
      }
    },
    "/api/v1/incidents/me": {
      "get": {
        "operationId": "CitizenIncidentsController_mine",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Citizen Incidents"
        ]
      }
    },
    "/api/v1/incidents/code/{caseCode}": {
      "get": {
        "operationId": "CitizenIncidentsController_code",
        "parameters": [
          {
            "name": "caseCode",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Citizen Incidents"
        ]
      }
    },
    "/api/v1/incidents/{id}": {
      "get": {
        "operationId": "CitizenIncidentsController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Citizen Incidents"
        ]
      }
    },
    "/api/v1/incidents/{id}/images": {
      "post": {
        "operationId": "CitizenIncidentsController_images",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Citizen Incidents"
        ]
      }
    },
    "/api/v1/admin/incidents": {
      "get": {
        "operationId": "AdminIncidentsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin Incidents"
        ]
      }
    },
    "/api/v1/admin/incidents/{id}": {
      "get": {
        "operationId": "AdminIncidentsController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin Incidents"
        ]
      }
    },
    "/api/v1/admin/incidents/{id}/status": {
      "patch": {
        "operationId": "AdminIncidentsController_status",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin Incidents"
        ]
      }
    },
    "/api/v1/admin/incidents/{id}/assignment": {
      "patch": {
        "operationId": "AdminIncidentsController_assignment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin Incidents"
        ]
      }
    },
    "/api/v1/admin/incidents/{id}/notes": {
      "post": {
        "operationId": "AdminIncidentsController_note",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin Incidents"
        ]
      }
    },
    "/api/v1/admin/dashboard/summary": {
      "get": {
        "operationId": "AdminOperationsController_summary",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin Operations"
        ]
      }
    },
    "/api/v1/admin/dashboard/recent-incidents": {
      "get": {
        "operationId": "AdminOperationsController_recent",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin Operations"
        ]
      }
    },
    "/api/v1/admin/dashboard/incidents-by-status": {
      "get": {
        "operationId": "AdminOperationsController_byStatus",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin Operations"
        ]
      }
    },
    "/api/v1/admin/dashboard/incidents-by-type": {
      "get": {
        "operationId": "AdminOperationsController_byType",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin Operations"
        ]
      }
    },
    "/api/v1/admin/users": {
      "get": {
        "operationId": "AdminOperationsController_users",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin Operations"
        ]
      },
      "post": {
        "operationId": "AdminOperationsController_createUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin Operations"
        ]
      }
    },
    "/api/v1/admin/users/{id}": {
      "get": {
        "operationId": "AdminOperationsController_user",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin Operations"
        ]
      },
      "patch": {
        "operationId": "AdminOperationsController_updateUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAdminUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin Operations"
        ]
      }
    },
    "/api/v1/admin/users/{id}/status": {
      "patch": {
        "operationId": "AdminOperationsController_userStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin Operations"
        ]
      }
    },
    "/api/v1/admin/users/{id}/reset-password": {
      "post": {
        "operationId": "AdminOperationsController_resetPassword",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin Operations"
        ]
      }
    },
    "/api/v1/admin/audit-logs": {
      "get": {
        "operationId": "AdminOperationsController_audit",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin Operations"
        ]
      }
    },
    "/api/v1/admin/settings": {
      "get": {
        "operationId": "AdminOperationsController_settings",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin Operations"
        ]
      },
      "patch": {
        "operationId": "AdminOperationsController_settingsUpdate",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin Operations"
        ]
      }
    },
    "/api/v1/notifications": {
      "get": {
        "operationId": "NotificationsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Notifications"
        ]
      }
    },
    "/api/v1/notifications/{id}/read": {
      "patch": {
        "operationId": "NotificationsController_read",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Notifications"
        ]
      }
    },
    "/api/v1/notifications/read-all": {
      "patch": {
        "operationId": "NotificationsController_readAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Notifications"
        ]
      }
    },
    "/api/v1/devices/register": {
      "post": {
        "operationId": "NotificationsController_device",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDeviceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Notifications"
        ]
      }
    },
    "/api/v1/health": {
      "get": {
        "operationId": "HealthController_health",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "System"
        ]
      }
    },
    "/api/v1/ready": {
      "get": {
        "operationId": "HealthController_ready",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "System"
        ]
      }
    }
  },
  "info": {
    "title": "Police Incident API",
    "description": "API ระบบแจ้งเหตุและติดตามสถานะเหตุฉุกเฉิน กองบินตำรวจ",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "FacebookLoginDto": {
        "type": "object",
        "properties": {}
      },
      "RefreshDto": {
        "type": "object",
        "properties": {}
      },
      "AdminLoginDto": {
        "type": "object",
        "properties": {}
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {}
      },
      "CreateIncidentDto": {
        "type": "object",
        "properties": {}
      },
      "StatusDto": {
        "type": "object",
        "properties": {}
      },
      "AssignmentDto": {
        "type": "object",
        "properties": {}
      },
      "NoteDto": {
        "type": "object",
        "properties": {}
      },
      "AdminUserDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateAdminUserDto": {
        "type": "object",
        "properties": {}
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {}
      },
      "RegisterDeviceDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}